requirements.txt
Flask
requests
numpy
Pillow
scikit-learn
gradio